+++ test/test_helper.exs
@@ -1,2 +1,2 @@
-K8s.Mock.DynamicHTTPProvider.start_link(nil)
+K8s.Client.DynamicHTTPProvider.start_link(nil)
 ExUnit.start()